refactor: change ambiguous `api' to `rest`
